fix(links): move header and footer links from talk to common settings
