+++ app/eventyay/eventyay_common/templates/eventyay_common/event/settings.html
@@ -58,6 +58,37 @@
             {% include "pretixcontrol/event/fragment_geodata.html" %}
             {% bootstrap_field sform.contact_mail layout="control" %}
             {% bootstrap_field sform.imprint_url layout="control" %}
+            <div class="form-group">
+                <label class="col-md-3 control-label">
+                    {% translate "Header links" %}
+                </label>
+                <div class="col-md-9">
+                    <div class="help-block text-muted">
+                        {% blocktranslate trimmed %}
+                            These links will be shown at the top of all your schedule-related pages, e.g.
+                            the schedule itself, speaker pages, session pages, etc.
+                            You could for example link to your home page, your registration page, or your
+                            livestream here.
+                        {% endblocktranslate %}
+                    </div>
+                    {% include "orga/includes/event_links_formset.html" with formset=header_links_formset %}
+                </div>
+            </div>
+
+            <div class="form-group">
+                <label class="col-md-3 control-label">
+                    {% translate "Footer links" %}
+                </label>
+                <div class="col-md-9">
+                    <div class="help-block text-muted">
+                        {% blocktranslate trimmed %}
+                            These links will be shown in the footer of all your public pages. You could
+                            for example link your terms of service, imprint, or privacy policy here.
+                        {% endblocktranslate %}
+                    </div>
+                    {% include "orga/includes/event_links_formset.html" with formset=footer_links_formset %}
+                </div>
+            </div>
             {% bootstrap_field form.is_public layout="control" %}
             {% comment %}
             <div class="setting-part">
